M1: residual SAC benchmark under model mismatch + README figure
Benchmark (benchmark/residual_pendulum/): Pendulum-v1 with plant mass
+40% vs the MPC's model. Residual SAC (full authority) starts near the
MPC baseline (-738 vs -629), never enters the catastrophic exploration
phase, and converges to -273; from-scratch SAC spends ~2.5k steps near
-1450 before converging to -236. Light/dark figures embedded in README.

Library changes surfaced by the benchmark:
- pendulum model: configurable m/l/g; smooth surrogate angle cost
  (gym's normalized-angle cost is kinked at the hanging position and
  breaks DDP)
- CrocoddylMPC: symmetry-breaking cold init (u_init), optional
  multistart; regression test for the hanging-start local minimum
- ResidualMPCEnv: float32 action space

+++ src/blendmpc/blends/residual.py
@@ -46,7 +46,7 @@ class ResidualMPCEnv(gym.Wrapper):
         self._high = np.asarray(env.action_space.high, dtype=float)
         self._scale = residual_scale * 0.5 * (self._high - self._low)
         self.action_space = gym.spaces.Box(
-            low=-1.0, high=1.0, shape=env.action_space.shape, dtype=np.float64
+            low=-1.0, high=1.0, shape=env.action_space.shape, dtype=np.float32
         )
         self._last_obs: np.ndarray | None = None
